Extract testing functionality
